requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
scikit-learn==1.4.0
numpy==1.26.4
gunicorn==21.2.0
